requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
python-dotenv
psycopg2
yfinance
os
pathlib
psycopg2-binary